\name{goldstandard.chr10}
\alias{goldstandard.chr10}
\docType{data}
\title{ 
  OCT4 mES Golden Standard(chromosome 10)
}
\description{
 0/1 results indicating whether the candidate sites are overlap with a ChIP-seq peaks. 
}
\usage{data(goldstandard.chr10)}
\format{
  The format is:
 num [1:10973] 0 0 0 0 0 0 0 0 0 0 ...
}
\examples{
data(goldstandard.chr10)
}
\keyword{datasets}
