\name{H1data.CpG.chr10}
\alias{H1data.CpG.chr10}
\docType{data}
\title{
 OCT4 H1 CpG Methylation(chr10)
}
\description{
  CpG methylation profile for OCT4 on H1-hESC cell line.
}
\usage{data(H1data.CpG.chr10)}
\format{
  A data frame with 7233 observations on the following 42 variables.
  \describe{
    \item{\code{methy.1}}{a numeric vector}
    \item{\code{methy.2}}{a numeric vector}
    \item{\code{methy.3}}{a numeric vector}
    \item{\code{methy.4}}{a numeric vector}
    \item{\code{methy.5}}{a numeric vector}
    \item{\code{methy.6}}{a numeric vector}
    \item{\code{methy.7}}{a numeric vector}
    \item{\code{methy.8}}{a numeric vector}
    \item{\code{methy.9}}{a numeric vector}
    \item{\code{methy.10}}{a numeric vector}
    \item{\code{methy.11}}{a numeric vector}
    \item{\code{methy.12}}{a numeric vector}
    \item{\code{methy.13}}{a numeric vector}
    \item{\code{methy.14}}{a numeric vector}
    \item{\code{methy.15}}{a numeric vector}
    \item{\code{methy.16}}{a numeric vector}
    \item{\code{methy.17}}{a numeric vector}
    \item{\code{methy.18}}{a numeric vector}
    \item{\code{methy.19}}{a numeric vector}
    \item{\code{methy.20}}{a numeric vector}
    \item{\code{methy.21}}{a numeric vector}
    \item{\code{total.1}}{a numeric vector}
    \item{\code{total.2}}{a numeric vector}
    \item{\code{total.3}}{a numeric vector}
    \item{\code{total.4}}{a numeric vector}
    \item{\code{total.5}}{a numeric vector}
    \item{\code{total.6}}{a numeric vector}
    \item{\code{total.7}}{a numeric vector}
    \item{\code{total.8}}{a numeric vector}
    \item{\code{total.9}}{a numeric vector}
    \item{\code{total.10}}{a numeric vector}
    \item{\code{total.11}}{a numeric vector}
    \item{\code{total.12}}{a numeric vector}
    \item{\code{total.13}}{a numeric vector}
    \item{\code{total.14}}{a numeric vector}
    \item{\code{total.15}}{a numeric vector}
    \item{\code{total.16}}{a numeric vector}
    \item{\code{total.17}}{a numeric vector}
    \item{\code{total.18}}{a numeric vector}
    \item{\code{total.19}}{a numeric vector}
    \item{\code{total.20}}{a numeric vector}
    \item{\code{total.21}}{a numeric vector}
  }
}
\examples{
data(H1data.CpG.chr10)
## maybe str(H1data.CpG.chr10) ; plot(H1data.CpG.chr10) ...
}
\keyword{datasets}
